+++ Mobile/app/(auth)/_layout.tsx
@@ -1,9 +1,12 @@
+import { StyleSheet, View } from "react-native";
 import { Tabs } from "expo-router";
 import TabBar from "../../components/Tab-Bar-component/tabbar";
+import Header from "../../components/Header/header";
 
 const AuthLayout = () => {
   return (
-    <>
+    <View style={styles.mainView}>
+      <Header />
       <Tabs
         screenOptions={{ headerShown: false }}
         tabBar={(props) => <TabBar {...props} />}
@@ -12,8 +15,15 @@ const AuthLayout = () => {
         <Tabs.Screen name="team" options={{ title: "Team" }} />
         <Tabs.Screen name="setting" options={{ title: "Setting" }} />
       </Tabs>
-    </>
+    </View>
   );
 };
 
 export default AuthLayout;
+
+const styles = StyleSheet.create({
+  mainView: {
+    flex: 1,
+    backgroundColor: "#001F3F",
+  },
+});
