+++ Mobile/components/Header/header.tsx
@@ -1,13 +1,13 @@
 import { StyleSheet, Text, View } from "react-native";
 import React from "react";
 import Logo from "./logo";
-import ProfileLink from "./profileLink";
+import HeaderNavigationLink from "./navigationLinks";
 
 const Header = () => {
   return (
     <View style={styles.mainView}>
       <Logo />
-      <ProfileLink />
+      <HeaderNavigationLink />
     </View>
   );
 };
